medguard-ai/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
PyMuPDF==1.24.3
Pillow==10.3.0
pytesseract==0.3.10
spacy==3.7.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
transformers==4.40.2
torch==2.2.2
numpy==1.26.4
python-docx==1.1.2
httpx==0.27.0
aiofiles==23.2.1
